refactor: couponPrice상태 끌어올리기 및 CouponModalContent 컴포넌트 역할에 맞게 props변경
